requirements.txt
flask==2.1.0
transformers==4.30.2
torch==1.13.1
Pillow==9.5.0
requests==2.28.2
python-dotenv==1.0.0
huggingface-hub==0.16.4